Started putting search into separate module.
