+++ b2handle/tests/handleclient_search_noaccess_test.py
@@ -7,6 +7,7 @@ else:
     import unittest
 import json
 sys.path.append("../..")
+from b2handle.searcher import Searcher
 from b2handle.handleclient import EUDATHandleClient
 from b2handle.handleexceptions import ReverseLookupException
 
@@ -20,6 +21,7 @@ class EUDATHandleClientSearchNoAccessTestCase(unittest.TestCase):
 
     def setUp(self):
         self.inst = EUDATHandleClient()
+        self.searcher = Searcher()
  
 
     def tearDown(self):
@@ -43,24 +45,24 @@ class EUDATHandleClientSearchNoAccessTestCase(unittest.TestCase):
 
     def test_create_revlookup_query_fulltext(self):
         with self.assertRaisesRegexp(ReverseLookupException, 'Full-text search is not implemented yet[.]*'):
-            self.inst.create_revlookup_query('foo', 'bar')
+            self.searcher.create_revlookup_query('foo', 'bar')
 
     def test_create_revlookup_query_forbiddenkeys(self):
         with self.assertRaisesRegexp(ReverseLookupException, 'Cannot search for key[.]*'):
-            self.inst.create_revlookup_query(foo='foo', bar='bar')
+            self.searcher.create_revlookup_query(foo='foo', bar='bar')
 
     def test_create_revlookup_query_noterms(self):
         with self.assertRaisesRegexp(ReverseLookupException, 'No search terms have been specified[.]*'):
-            self.inst.create_revlookup_query()
+            self.searcher.create_revlookup_query()
 
     def test_create_revlookup_query_norestriction(self):
-        inst = EUDATHandleClient(allowed_search_keys=[])
-        query = inst.create_revlookup_query(baz='baz')
+        searcher = Searcher(allowed_search_keys=[])
+        query = searcher.create_revlookup_query(baz='baz')
         self.assertEqual(query, '?baz=baz',
             'The query is: '+query)
 
     def test_create_revlookup_query_normal(self):
-        query = self.inst.create_revlookup_query(URL='foo')
+        query = self.searcher.create_revlookup_query(URL='foo')
         self.assertEqual(query, '?URL=foo',
             'The query is: '+query)
